docs: 디렉토리 구조 수정, 파일 분할
